Move ContextualPath and PermissionBadge from Valhalla
