+++ app/views/catalog/_terms_ephemera_vocabulary.html.erb
@@ -1,6 +1,6 @@
 <h2>Terms</h2>
 <ul>
   <% decorated_resource.terms.each do |term| %>
-    <li><%= link_to term.label, Valhalla::ContextualPath.new(child: term).show %></li>
+    <li><%= link_to term.label, ContextualPath.new(child: term).show %></li>
   <% end %>
 </ul>
