+++ spec/shared_specs.rb
@@ -3,6 +3,6 @@ require_relative "shared_specs/controllers/base_resource_controller.rb"
 require_relative "shared_specs/decorators/collection_decorator.rb"
 require_relative "shared_specs/models/linked_data/linked_resource.rb"
 require_relative "shared_specs/services/event_generator.rb"
-require_relative "shared_specs/valhalla/change_sets/change_set.rb"
+require_relative "shared_specs/change_sets/change_set.rb"
 require_relative "shared_specs/valhalla/models/resource.rb"
 require_relative "shared_specs/valkyrie/decorators/resource_decorator.rb"
