--- valhalla/app/helpers/valhalla/application_helper.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-module Valhalla
-  module ApplicationHelper
-    def visibility_badge(value)
-      Valhalla::PermissionBadge.new(value).render
-    end
-  end
-end
