--- valhalla/app/helpers/valhalla/contextual_path_helper.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-module Valhalla
-  module ContextualPathHelper
-    def contextual_path(child, parent)
-      ContextualPath.new(child: child, parent_id: parent.try(:id))
-    end
-  end
-end
